Don't send no-op updates
Avoids sending a resource update that has no actual changes compared to the last value sent.

This can happen because of how we process services/endpoints. When they are updated, we attempt to apply that updated data to existing DCP containers/executables. Sometimes this results in no actual difference, so we can drop these updates and avoid sending them over the wite and triggering UI refreshes.

As an example, when loading dashboard for the eShopLite sample solution, I see up to six redundant updates that are now avoided.